[Fix] set message page 컴포넌트 분리
